refact: api 공통화
