--- packages/web-domains/src/common/apis/schema/HobbyListResponse.ts
@@ -1,8 +1,0 @@
-export type HobbyType = {
-  hobbyId: number;
-  content: string;
-};
-
-export type HobbyListResponse = {
-  contents: HobbyType[];
-};
